+++ lib/modules/test/test_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: test
-// Updated at: 2026-06-25 16:57:51
-// Commit: docs(test): update readme and API contracts for test
+// Updated at: 2026-06-25 19:14:35
+// Commit: fix(test): fix memory leak by cancelling test subscription
 
 class testComponentState {
-  final String commitRef = '2311';
-  final String lastMessage = 'docs(test): update readme and API contracts for test';
-  final DateTime timestamp = DateTime.parse('2026-06-25T16:57:51+05:00');
+  final String commitRef = '2326';
+  final String lastMessage = 'fix(test): fix memory leak by cancelling test subscription';
+  final DateTime timestamp = DateTime.parse('2026-06-25T19:14:35+05:00');
 }